+++ Pulp/Common/Memory/AllocationPolicies/StackAllocator.cpp
@@ -20,7 +20,7 @@ StackAllocator::StackAllocator(void* start, void* end) :
 #if X_ENABLE_MEMORY_ALLOCATOR_STATISTICS
 	zero_this( &m_statistics );
 
-	int Len = safe_static_cast<UINT,int>( (UINT)end - (UINT)start );
+	size_t Len = safe_static_cast<size_t, uintptr_t>((uintptr_t)end - (uintptr_t)start);
 
 	m_statistics.m_type = "Stack";
 	m_statistics.m_virtualMemoryReserved = Len;
@@ -57,7 +57,7 @@ void* StackAllocator::allocate( size_t size, size_t alignment, size_t align_offs
 	m_statistics.m_allocationCount++;
 	m_statistics.m_internalOverhead += sizeof( BlockHeader );
 	m_statistics.m_physicalMemoryUsed += size + (m_current - old_current);
-	m_statistics.m_wasteAlignment += safe_static_cast<uint32_t>( ((uint32_t)m_current - (uint32_t)m_start) - allocationOffset );
+	m_statistics.m_wasteAlignment += safe_static_cast<size_t>(((uintptr_t)m_current - (uintptr_t)m_start) - allocationOffset);
 
 	m_statistics.m_wasteAlignmentMax = Max( m_statistics.m_wasteAlignment, m_statistics.m_wasteAlignmentMax );
 	m_statistics.m_allocationCountMax = Max( m_statistics.m_allocationCount, m_statistics.m_allocationCountMax );
